Move HomeActivity to app package.
