+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name=".home.HomeActivity"/>
+        <activity android:name=".app.home.HomeActivity"/>
         <activity
             android:name=".auth.LoginActivity"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar"/>
