+++ app/src/main/java/com/bsstokes/bspix/app/home/HomeActivity.java
@@ -1,4 +1,4 @@
-package com.bsstokes.bspix.home;
+package com.bsstokes.bspix.app.home;
 
 import android.content.Context;
 import android.content.Intent;
