+++ app/src/main/java/com/bsstokes/bspix/di/AppComponent.java
@@ -2,7 +2,7 @@ package com.bsstokes.bspix.di;
 
 import com.bsstokes.bspix.BsPixApplication;
 import com.bsstokes.bspix.auth.LoginActivity;
-import com.bsstokes.bspix.home.HomeActivity;
+import com.bsstokes.bspix.app.home.HomeActivity;
 import com.bsstokes.bspix.main.MainActivity;
 
 public interface AppComponent {
